### 2024-08-09 #7077 @bmcfee
Back at a proper keyboard now and not trying to wrangle github responses with a toddler in hand. :)  Reviewer invitations are flying through the ether now.

@josemanuel22 while we're working on finding reviewers, it would be helpful if you could fix the references section for the paper.  Reviewers will ask for / need this anyway, so we may as well get out ahead of it.
