+++ paper.md
@@ -27,16 +27,16 @@ Implicit generative models have the capability to learn arbitrary complex data d
 
 In addition to its core functionalities, `ISL` offers a suite of utility functions, including support for generative adversarial networks and tools for time series analysis and generation. This makes `ISL` a valuable resource for both machine learning researchers and data scientists/software developers who seek to train their models with this novel approach.
 
-`ISL` has already contributed in a scientific publications [de2024training], underscoring its utility and impact in the field. Its combination of speed, thoughtful design, and robust machine learning functionalities for Implicit Generative Models positions `ISL` as a powerful tool for advancing scientific research and practical applications alike in the area of Implicit Generative Models.
+`ISL` has already contributed in a scientific publications [@de2024training], underscoring its utility and impact in the field. Its combination of speed, thoughtful design, and robust machine learning functionalities for Implicit Generative Models positions `ISL` as a powerful tool for advancing scientific research and practical applications alike in the area of Implicit Generative Models.
 
 # Methods
 
 Implicit generative models employ an $m$-dimensional latent random variable (r.v.) $\mathbf{z}$ to simulate random samples from a prescribed $n$-dimensional target probability distribution. To be precise, the latent variable undergoes a transformation through a deterministic function $g_{\theta}$, which maps $\mathbb{R}^m \mapsto \mathbb{R}^n$ using the parameter set $\theta$. Given the model capability to generate samples with ease, various techniques can be employed for contrasting two sample collections: one originating from the genuine data distribution and the other from the model distribution. This approach essentially constitutes a methodology for the approximation of probability distributions via comparison.
-Generative adversarial networks (GANs) [goodfellow2014generative], $f$-GANs [nowozin2016f], Wasserstein-GANs (WGANs) [arjovsky2017wasserstein], adversarial variational Bayes (AVB) [mescheder2017adversarial], and maximum mean-miscrepancy (MMD) GANs [li2017mmd] are some popular methods that fall within this framework. 
+Generative adversarial networks (GANs) [@goodfellow2014generative], $f$-GANs [@nowozin2016f], Wasserstein-GANs (WGANs) [@arjovsky2017wasserstein], adversarial variational Bayes (AVB) [@mescheder2017adversarial], and maximum mean-miscrepancy (MMD) GANs [li2017mmd] are some popular methods that fall within this framework. 
 
-Approximation of 1-dimensional (1D) parametric distributions is a seemingly naive problem for which the above-mentioned models can perform below expectations. In [zaheer2017gan], the authors report that various types of GANs struggle to approximate relatively simple distributions from samples, emerging with MMD-GAN as the most promising technique. However, the latter implements a kernelized extension of a moment-matching criterion defined over a reproducing kernel Hilbert space, and consequently, the objective function is expensive to compute. 
+Approximation of 1-dimensional (1D) parametric distributions is a seemingly naive problem for which the above-mentioned models can perform below expectations. In [@zaheer2017gan], the authors report that various types of GANs struggle to approximate relatively simple distributions from samples, emerging with MMD-GAN as the most promising technique. However, the latter implements a kernelized extension of a moment-matching criterion defined over a reproducing kernel Hilbert space, and consequently, the objective function is expensive to compute. 
 
[…]
