### 2024-09-02 #7077 @josemanuel22
@bmcfee I'm sorry to bother you; this is the first time I've submitted something to JOSS. What is the current status of this?
