### 2024-09-23 #7272 @editorialbot
[REVIEW]: ISL: A Julia package for training Implicit Generative Models via an Invariant Statistical Loss
**Submitting author:** @https://github.com/josemanuel22 (<a href="http://orcid.org/0009-0002-5251-0515">José Manuel de Frutos</a>)
**Repository:** https://github.com/josemanuel22/ISL
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
