### 2024-10-08 #7272 @odunbar
Hi folks, just working on the review here, I had a question about plagarism,
- The author (with collaborators) have published this PMLR paper https://proceedings.mlr.press/v238/frutos24a.html
- The summary and statement of need in the JOSS submission are verbatim copies of the abstract and introduction of this paper.

Despite this being the authors own work, I assume this counts as plagarism and should be changed sufficiently in wording for the JOSS publication? Or is this "OK" here?
