### 2024-10-11 #7272 @bmcfee
Thanks @odunbar for raising this.  I think the relevant part of the submission guidelines is [here](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software):

> Co-publication of science, methods, and software

> Sometimes authors prepare a JOSS publication alongside a contribution describing a science application, details of algorithm development, and/or methods assessment. In this circumstance, JOSS considers submissions for which the implementation of the software itself reflects a substantial scientific effort. This may be represented by the design of the software, the implementation of the algorithms, creation of tutorials, or any other aspect of the software. We ask that authors indicate whether related publications (published, in review, or nearing submission) exist as part of submitting to JOSS.

In this case, the related paper is cited in the JOSS submission, but there was not a separate explicit mention of the work in the submission thread.

I've bumped this up to @openjournals/joss-eics for comment, and will get back to this shortly.
