### 2024-10-11 #7272 @bmcfee
@josemanuel22 After discussing with EiCs, we agree that the verbatim text in the summary and statement of need would constitute self-plagiarism and should be rewritten.  Can you take care of this, and respond back in this thread?
